[web] Move InstallButton to its own file
