+++ web/src/Overview.jsx
@@ -20,11 +20,10 @@
  */
 
 import React, { useState } from "react";
-import { useInstallerClient } from "./context/installer";
 import { useSoftware } from "./context/software";
 import { useNavigate, Navigate } from "react-router-dom";
 
-import { Button, Flex, FlexItem, Text } from "@patternfly/react-core";
+import { Button, Flex, FlexItem } from "@patternfly/react-core";
 
 import { Title, PageIcon, PageActions, MainActions } from "./Layout";
 import Category from "./Category";
@@ -32,7 +31,7 @@ import LanguageSelector from "./LanguageSelector";
 import Storage from "./Storage";
 import Users from "./Users";
 import Network from "./Network";
-import Popup from "./Popup";
+import InstallButton from "./InstallButton";
 
 import {
   EOS_SOFTWARE as OverviewIcon,
@@ -59,76 +58,6 @@ const ChangeProductButton = () => {
       aria-label="Change selected product"
       onClick={() => navigate("/products")}
     />
-  );
-};
-
-const InstallConfirmationPopup = ({ onAccept, onClose }) => (
-  <Popup
-    title="Confirm Installation"
-    isOpen="true"
-  >
-    <Text>
-      If you continue, partitions on your hard disk will be modified according to the
-      installation settings in the previous dialog.
-    </Text>
-    <Text>
-      Please, cancel and check the settings if you are unsure.
-    </Text>
-
-    <Popup.Actions>
-      <Popup.Confirm onClick={onAccept}>Install</Popup.Confirm>
-      <Popup.Cancel onClick={onClose} autoFocus />
-    </Popup.Actions>
-  </Popup>
-);
-
-const CannotInstallPopup = ({ onClose }) => (
-  <Popup
-    title="Problems Found"
-    isOpen="true"
-  >
-    <Text>
-      Some problems were found when trying to start the installation.
-      Please, have a look to the reported issues and try again after.
-    </Text>
-
-    <Popup.Actions>
-      <Popup.Cancel onClick={onClose} autoFocus />
-    </Popup.Actions>
-  </Popup>
-);
-
-const renderPopup = (error, { onAccept, onClose }) => {
-  if (error) {
-    return <CannotInstallPopup onClose={onClose} />;
-  } else {
-    return <InstallConfirmationPopup onClose={onClose} onAccept={onAccept} />;
-  }
-};
-
-const InstallButton = ({ onClick }) => {
-  const client = useInstallerClient();
-  const [isOpen, setIsOpen] = useState(false);
-  const [error, setError] = useState(false);
-
-  const open = () => {
-    onClick();
-    client.manager.canInstall().then(ok => {
-      setIsOpen(true);
-      setError(!ok);
-    });
-  };
-  const close = () => setIsOpen(false);
-  const install = () => client.manager.startInstallation();
-
-  return (
-    <>
-      <Button isLarge variant="primary" onClick={open}>
-        Install
-      </Button>
-
-      { isOpen && renderPopup(error, { onAccept: install, onClose: close }) }
-    </>
   );
 };
